Moved keyframe test to keyframes.test.tsx
